+++ .eslintrc.js
@@ -1,2 +1,0 @@
-const config = require('./index')
-module.exports = config