+++ index.js
@@ -17,9 +17,7 @@ module.exports = {
   },
 
   'plugins': [
-    'ava',
-    'import',
-    'react'
+    'import'
   ],
 
   'rules': {
@@ -315,29 +313,6 @@ module.exports = {
     'yield-star-spacing': ['error', 'after'],
     'yoda': 'error',
 
-    'ava/assertion-arguments': ['error', {
-      'message': 'never'
-    }],
-    'ava/assertion-message': ['error', 'never'],
-    'ava/max-asserts': ['error', 1],
-    'ava/no-cb-test': 'error',
-    'ava/no-identical-title': 'error',
-    'ava/no-ignored-test-files': 'off',
-    'ava/no-invalid-end': 'error',
-    'ava/no-only-test': 'warn',
-    'ava/no-skip-assert': 'warn',
-    'ava/no-skip-test': 'warn',
-    'ava/no-statement-after-end': 'error',
-    'ava/no-todo-test': 'warn',
-    'ava/no-unknown-modifiers': 'error',
-    'ava/prefer-power-assert': 'off',
-    'ava/test-ended': 'error',
-    'ava/test-title': ['error', 'if-multiple'],
-    'ava/use-t': 'error',
-    'ava/use-t-well': 'error',
-    'ava/use-test': 'error',
-    'ava/use-true-false': 'off',
-
     'import/default': 'error',
     'import/export': 'error',
     'import/extensions': ['error', {
@@ -361,63 +336,6 @@ module.exports = {
     'import/no-unresolved': 'error',
     'import/order': ['error', {
       'groups': ['builtin', 'external', 'internal', 'parent', 'sibling', 'index']
-    }],
-
-    'react/display-name': 'off',
-    'react/forbid-prop-types': 'off',
-    'react/jsx-boolean-value': 'error',
-    'react/jsx-closing-bracket-location': 'error',
-    'react/jsx-curly-spacing': ['error', 'never'],
-    'react/jsx-equals-spacing': ['error', 'never'],
-    'react/jsx-first-prop-new-line': ['error', 'multiline'],
-    'react/jsx-handler-names': ['error', {
-      'eventHandlerPrefix': 'handle',
-      'eventHandlerPropPrefix': 'on'
-    }],
-    'react/jsx-indent': ['error', 2],
-    'react/jsx-indent-props': ['error', 2],
-    'react/jsx-key': 'error',
-    'react/jsx-max-props-per-line': ['error', {
-      'maximum': 3
-    }],
-    'react/jsx-no-bind': ['error', {
-      'ignoreRefs': true,
-      'allowArrowFunctions': true,
-      'allowBind': false
-    }],
-    'react/jsx-no-duplicate-props': ['error', {
-      'ignoreCase': true
-    }],
-    'react/jsx-no-literals': 'off',
-    'react/jsx-no-undef': 'error',
-    'react/jsx-pascal-case': 'error',
-    'react/jsx-sort-props': 'off',
-    'react/jsx-space-before-closing': ['error', 'always'],
-    'react/jsx-uses-react': 'error',
-    'react/jsx-uses-vars': 'error',
-    'react/no-danger': 'error',
-    'react/no-deprecated': 'error',
-    'react/no-did-mount-set-state': 'off',
-    'react/no-did-update-set-state': 'off',
-    'react/no-direct-mutation-state': 'error',
-    'react/no-is-mounted': 'error',
-    'react/no-multi-comp': ['error', {
-      'ignoreStateless': true
-    }],
-    'react/no-set-state': 'off',
-    'react/no-string-refs': 'error',
-    'react/no-unknown-property': 'error',
-    'react/prefer-es6-class': ['error', 'always'],
-    'react/prefer-stateless-function': 'error',
-    'react/prop-types': 'off',
-    'react/react-in-jsx-scope': 'warn',
-    'react/require-extension': ['warn', {
-      'extensions': ['.js', '.jsx']
-    }],
-    'react/require-render-return': 'error',
-    'react/self-closing-comp': 'error',
-    'react/sort-comp': 'off',
-    'react/sort-prop-types': 'off',
-    'react/wrap-multilines': 'error'
+    }]
   }
 }