+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.3.0]
+## [0.4.0] - 2016-05-07
+### Changed
+- Move AVA and React rules to individual config files
+- Extend usage guide in README
+
+## [0.3.0] - 2016-05-07
 ### Added
 - New rules
 - React specifix rules
@@ -37,6 +42,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.1.0 - 2016-04-16
 
+[0.3.0]: https://github.com/sadorlovsky/codestyle/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/sadorlovsky/codestyle/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/sadorlovsky/codestyle/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/sadorlovsky/codestyle/compare/v0.2.0...v0.2.1